Move App.js out of components folder and fix test bug
